requirements.txt
asgiref==3.8.1
Django==5.1.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
phonenumbers==8.13.47
sqlparse==0.5.1
tzdata==2024.2
